+++ Assets/Tests/Behavior/ServiceInterfaces/LateTickableTests.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections;
 using Autofac;
 using MrWatts.Internal.FuelInject.Testing;
+using MrWatts.Internal.FuelInject.Testing.Utility;
 using NUnit.Framework;
-using UnityEngine;
 using UnityEngine.TestTools;
 
 namespace MrWatts.Internal.FuelInject.TestProject.Tests.Behaviour
@@ -26,9 +26,7 @@ namespace MrWatts.Internal.FuelInject.TestProject.Tests.Behaviour
                 }
             );
 
-            yield return new WaitForEndOfFrame();
-
-            Assert.IsTrue(wasCalled);
+            yield return new WaitTemporarilyUntil(() => wasCalled);
         }
 
         [UnityTest]
@@ -50,9 +48,7 @@ namespace MrWatts.Internal.FuelInject.TestProject.Tests.Behaviour
             Assert.IsFalse(listeningUnityKernelLogger.WasExceptionLogged);
 
             yield return sceneLoadingCoroutine;
-            yield return new WaitForEndOfFrame();
-
-            Assert.IsTrue(listeningUnityKernelLogger.WasExceptionLogged);
+            yield return new WaitTemporarilyUntil(() => listeningUnityKernelLogger.WasExceptionLogged);
         }
     }
 }